% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{ar_acf}
\alias{ar_acf}
\title{Compute the autocovariance function of an AR process.}
\usage{
ar_acf(n, phis, sd)
}
\arguments{
\item{n}{The number of lags for which to compute the ACF.}

\item{phis}{A numeric vector of AR coefficients.}

\item{sd}{The standard deviation driving the process.}
}
\value{
A numeric vector representing the ACF of the AR process.
}
\description{
This function computes the autocovariance function (ACF) of an autoregressive
(AR) process with given AR coefficients \code{phis} and standard deviation
\code{sd}. It is an add on to the default R code that only computes the
autocorrelation function.
}
\examples{
ar_acf(5, c(0.5, -0.3), 1)

}
