% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.r
\name{ar_spectrum}
\alias{ar_spectrum}
\title{Compute the autoregressive (AR) spectrum.}
\usage{
ar_spectrum(ff, phis, sd, delta = 1)
}
\arguments{
\item{ff}{A numeric vector of frequencies.}

\item{phis}{A numeric vector of AR coefficients.}

\item{sd}{The standard deviation.}

\item{delta}{The sampling interval (default is 1).}
}
\value{
A numeric vector representing the AR spectrum.
}
\description{
This function computes the autoregressive (AR) spectrum given the frequencies
\code{ff}, the AR coefficients \code{phis}, the standard deviation \code{sd},
and the sampling interval \code{delta}.
}
\examples{
ff <- speccy::get_ff(1000)
phis <- c(0.5, -0.3)
sd <- 1
ar_spectrum(ff, phis, sd)

}
