% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bases.R
\name{build_bases}
\alias{build_bases}
\title{Build biased bases debiased spectral analysis}
\usage{
build_bases(
  n,
  delta = 1,
  h = NULL,
  k = NULL,
  centres = NULL,
  widths = NULL,
  lowers = NULL,
  uppers = NULL,
  lag_sequence = NULL,
  ...
)
}
\arguments{
\item{n}{The length of the input signal.}

\item{delta}{The sampling interval (default is 1).}

\item{h}{A numeric vector taper weights to apply to the time series (default
is NULL, equal weighting is assumed).}

\item{k}{The number of bases to build with evenly spaced centres and widths
(default is NULL).}

\item{centres}{A numeric vector representing the centres of the bases
(default is NULL).}

\item{widths}{A numeric vector representing the widths of the bases
(default is NULL).}

\item{lowers}{A numeric vector representing the lower bounds of the bases
(default is NULL).}

\item{uppers}{A numeric vector representing the upper bounds of the bases
(default is NULL).}

\item{...}{Frequency subsetting options (all default to TRUE).
\code{incl_boundaries} Logical indicating whether to include the boundary
frequencies at zero and Nyquist. \code{one_sided} Logical indicating whether
to generate a one-sided frequency vector. \code{positive_freqs} Logical
indicating whether to include results as positive frequencies or
zero-centred.}
}
\value{
A matrix of bases with dimensions \code{n} by \code{k}.
}
\description{
This function biased bases for debiased spectral analysis
based on the specified parameters. The bases can be constructed using
different methods:
\itemize{
\item If \code{k} is provided, it generates bases with evenly spaced centres
and widths.
\item If \code{centres} and \code{widths} are provided, it uses the specified
centres and widths for the bases.
\item If \code{lowers} and \code{uppers} are provided, it constructs bases with
bounded centres and widths.
}
}
\examples{
build_bases(100, k = 5)

}
