% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convolve_taper}
\alias{convolve_taper}
\title{Computes an open convolution of a taper with itself.}
\usage{
convolve_taper(h)
}
\arguments{
\item{h}{A numeric vector of taper weights.}
}
\value{
A numeric vector representing the tapered convolution result.
}
\description{
This function computes open convolution of a taper \code{h} with itself
and normalises.
}
\examples{
h <- c(1, 2, 3, 2, 1)
convolve_taper(h)

}
