% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bases.R
\name{get_centres}
\alias{get_centres}
\title{Compute the centres of basis bins.}
\usage{
get_centres(k, delta = 1)
}
\arguments{
\item{k}{The number of bases.}

\item{delta}{The sampling interval (default is 1).}
}
\value{
A list containing the centres of the frequency bins and the width of
each bin.
}
\description{
This function computes the centres of \code{k} bases, each of width
\code{width}.
}
\examples{
get_centres(10, 0.1)

}
