% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_ff}
\alias{get_ff}
\title{Generate a frequency vector for FFT.}
\usage{
get_ff(
  n,
  delta = 1,
  incl_boundaries = TRUE,
  one_sided = TRUE,
  positive_freqs = TRUE
)
}
\arguments{
\item{n}{The length of the input signal.}

\item{delta}{The sampling interval (default is 1).}

\item{incl_boundaries}{Logical indicating whether to include the boundary
frequencies at zero and Nyquist (default is TRUE).}

\item{one_sided}{Logical indicating whether to generate a one-sided frequency
vector (default is TRUE).}

\item{positive_freqs}{Logical indicating whether to include only positive
frequencies (default is TRUE).}
}
\value{
A numeric vector representing the frequency vector for FFT.
}
\description{
This function generates a frequency vector for FFT with given parameters.
}
\examples{
get_ff(10)

}
