% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_window.R
\name{lag_window}
\alias{lag_window}
\title{Compute the lag window estimate.}
\usage{
lag_window(
  ts,
  h = NULL,
  delta = 1,
  design_kernel = NULL,
  lag_sequence = NULL,
  return_ff = TRUE,
  acf = FALSE,
  incl_boundaries = TRUE,
  one_sided = TRUE,
  positive_freqs = TRUE
)
}
\arguments{
\item{ts}{A numeric vector representing the time series.}

\item{h}{A numeric vector taper weights to apply to the time series (default
is NULL, equal weighting is assumed).}

\item{delta}{The sampling interval (default is 1).}

\item{design_kernel}{A numeric vector representing the non-negative side of
the design kernel (default is NULL). The first value is taken as the zeroth
value, and the remainder is the positive side of the kernel.}

\item{lag_sequence}{A numeric vector representing the lag sequence (default
is NULL).}

\item{return_ff}{Logical indicating whether to return the frequency vector
(default is TRUE).}

\item{acf}{Logical indicating whether to treat ts as the ACF. Setting to TRUE
returns the expected value of the multitaper estimate (default is FALSE).}

\item{incl_boundaries}{Logical indicating whether to include the boundary
frequencies at zero and Nyquist (default is TRUE).}

\item{one_sided}{Logical indicating whether to generate a one-sided frequency
vector (default is TRUE).}

\item{positive_freqs}{Logical indicating whether to include results as
positive frequencies or zero-centred (default is TRUE).}
}
\value{
A list with components \code{ff} (the frequency vector) and
\code{estimate} (the lag window estimate).
}
\description{
This function computes the lag window estimate using either a design kernel
or a lag sequence.
}
