% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_window.R
\name{make_design_window}
\alias{make_design_window}
\title{Create a design window for lag window estimators.}
\usage{
make_design_window(x, n)
}
\arguments{
\item{x}{A numeric vector representing the non-negative design kernel.}

\item{n}{Length of the output design window.}
}
\value{
A normalized design window.
}
\description{
This function creates a design window for lag window estimation given the
non-negative values. The first value of \code{x} is the zeroth value, and the
remainder is the positive side of the window. The design window is returned
for a length \code{n} vector with the zero-frequency listed first.
}
\examples{
make_design_window(rep(1, 32), 1028)

}
