% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitaper.R
\name{multitaper}
\alias{multitaper}
\title{Compute the multitaper spectral estimate.}
\usage{
multitaper(ts, tapers, delta = 1, return_ff = TRUE, acf = FALSE)
}
\arguments{
\item{ts}{A numeric vector representing the time series.}

\item{tapers}{A matrix of tapers (each column is a taper).}

\item{delta}{The spacing between frequency bins (default is 1).}

\item{return_ff}{Logical indicating whether to return the frequency vector (default is TRUE).}

\item{acf}{Logical indicating whether to compute the autocovariance function (default is FALSE).}
}
\value{
A list with components \code{ff} (the frequency vector) and \code{mt} (the multitaper spectral estimate).
}
\description{
This function computes the multitaper spectral estimate of a time series \code{ts}
using the given set of tapers. The multitaper method uses multiple tapers to
reduce variance in the spectral estimate.
}
\examples{
ts <- sin(seq(0, 10, by = 0.1))
tapers <- matrix(data = rnorm(100), nrow = 100, ncol = 5)
multitaper(ts, tapers)

}
