% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{shift_index}
\alias{shift_index}
\title{Calculate shift index for Welch's estimator}
\usage{
shift_index(l, p)
}
\arguments{
\item{l}{An integer representing the index to shift.}

\item{p}{A numeric proportion by which to shift the index.}
}
\value{
The shift index \code{s}.
}
\description{
This function shifts an index \code{l} by a proportion \code{p}.
}
\examples{
shift_index(10, 0.5)

}
