% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handy_functions.R
\name{sinc}
\alias{sinc}
\title{Sinc function.}
\usage{
sinc(x, normalise = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{normalise}{Logical indicating whether to normalize \code{x} by
multiplying by pi (default is FALSE).}
}
\value{
A numeric vector representing the sinc function values.
}
\description{
This function computes the sinc function for a given input \code{x}. The sinc
function is defined as sin(x) / x, with special handling for x = 0 where the
value is defined as 1.
}
\examples{
x <- seq(-10, 10, length.out = 100)
sinc(x)

}
