% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_locations}
\alias{subset_locations}
\title{Subset frequency locations based on specified criteria.}
\usage{
subset_locations(
  x,
  incl_boundaries = TRUE,
  one_sided = TRUE,
  positive_freqs = TRUE
)
}
\arguments{
\item{x}{A numeric vector with default R frequency orderings.}

\item{incl_boundaries}{Logical indicating whether to include the boundary
frequencies at zero and Nyquist (default is TRUE).}

\item{one_sided}{Logical indicating whether to generate a one-sided frequency
vector (default is TRUE).}

\item{positive_freqs}{Logical indicating whether to include results as
positive frequencies or zero-centred (default is TRUE).}
}
\value{
A numeric vector of subsetted frequency orderings.
}
\description{
This function subsets frequency locations based on the specified criteria,
such as whether to include boundaries, whether to consider only positive
frequencies, and whether to return a one-sided frequency vector.
}
\examples{
x <- seq(-0.5, 0.5, by = 0.1)
subset_locations(x)

}
