% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_sequences.R
\name{w_bartlett}
\alias{w_bartlett}
\title{Generate a Bartlett sequence.}
\usage{
w_bartlett(n, l)
}
\arguments{
\item{n}{The length of the Bartlett window.}

\item{l}{The parameter for the Bartlett window.}
}
\value{
A Bartlett window of length \code{n}.
}
\description{
This function generates Bartlett's sequence of length \code{n} with parameter
\code{l}.
}
\examples{
w_bartlett(10, 5)

}
