% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_sequences.R
\name{w_daniell}
\alias{w_daniell}
\title{Generate a Daniell sequence.}
\usage{
w_daniell(n, l)
}
\arguments{
\item{n}{The length of the Daniell sequence.}

\item{l}{The parameter for the Daniell sequence.}
}
\value{
A Daniell window of length \code{n}.
}
\description{
This function generates Daniell's sequence of length \code{n} with parameter
\code{l}.
}
\examples{
w_daniell(10, 5)

}
