% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{welch}
\alias{welch}
\title{Compute the Welch's spectral estimate.}
\usage{
welch(
  ts,
  m,
  l,
  s = NULL,
  overlap = NULL,
  delta = 1,
  h = NULL,
  return_ff = TRUE,
  ...
)
}
\arguments{
\item{ts}{A numeric vector representing the time series.}

\item{m}{The number of segments to divide the time series into.}

\item{l}{The length of each segment.}

\item{s}{The overlap between segments (default is NULL).}

\item{overlap}{The overlap between segments as a proportion (default is
NULL).}

\item{delta}{The sampling interval (default is 1).}

\item{h}{A numeric vector taper weights to apply to the time series (default
is NULL, equal weighting is assumed).}

\item{return_ff}{Logical indicating whether to return the frequency vector
(default is TRUE).}

\item{...}{Frequency subsetting options (all default to TRUE).
\code{incl_boundaries} Logical indicating whether to include the boundary
frequencies at zero and Nyquist. \code{one_sided} Logical indicating whether
to generate a one-sided frequency vector. \code{positive_freqs} Logical
indicating whether to include results as positive frequencies or
zero-centred.}
}
\value{
A list with components \code{ff} (the frequency vector) and
\code{pwelch} (the Welch's periodogram estimate).
}
\description{
This function computes Welch's spectral estimate of a time series \code{ts}
using the specified parameters \code{m}, \code{l}, and \code{s}.
}
