% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{welch_length}
\alias{welch_length}
\title{Calculate the length of time series from Welch parameters.}
\usage{
welch_length(m, l, p = NULL, s = NULL)
}
\arguments{
\item{m}{The number of segments.}

\item{l}{The length of each segment.}

\item{p}{The proportion of overlap between segments (default is NULL).}

\item{s}{The shift value between segments (default is NULL).}
}
\value{
The length of a Welch segment.
}
\description{
This function calculates the length of a time series specified by the
parameterisation of Welch's method.
}
\examples{
welch_length(5, 10, p = 0.5)

}
