% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bochner.R
\name{bochner}
\alias{bochner}
\title{Compute the PSD from an ACF using Bochner's theorem.}
\usage{
bochner(
  acf,
  h = NULL,
  delta = 1,
  bias = TRUE,
  return_ff = TRUE,
  incl_boundaries = TRUE,
  one_sided = TRUE,
  positive_freqs = TRUE
)
}
\arguments{
\item{acf}{A numeric vector representing the autocovariance function.}

\item{bias}{}

\item{full}{Logical indicating whether to return the full periodogram or only the positive frequencies (default is FALSE).}

\item{zero}{Logical indicating whether to return the zero-frequency component (default is FALSE).}
}
\value{
A numeric vector representing the periodogram.
}
\description{
This function computes the PSD using Bochner's (or Wiener-Khinchin's)
theorem.
, which involves modifying the autocovariance function (ACF) and then taking the Fourier transform of the modified ACF.
}
\examples{
acf <- c(1, 0.5, 0.3, 0.1, 0, -0.1, -0.2)
bochner(acf)

}
