% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ff}
\alias{get_ff}
\title{Generate a frequency vector for FFT.}
\usage{
get_ff(n, delta = 1, ...)
}
\arguments{
\item{n}{The length of the input signal.}

\item{delta}{The sampling interval (default is 1).}

\item{...}{Frequency subsetting options (all default to TRUE).
\code{incl_boundaries} Logical indicating whether to include the boundary
frequencies at zero and Nyquist. \code{one_sided} Logical indicating whether
to generate a one-sided frequency vector. \code{positive_freqs} Logical
indicating whether to include results as positive frequencies or
zero-centred.}
}
\value{
A numeric vector representing the frequency vector for FFT.
}
\description{
This function generates a frequency vector for FFT with given parameters.
}
\examples{
get_ff(10)

}
