% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitaper.R
\name{multitaper}
\alias{multitaper}
\title{Compute the multitaper spectral estimate.}
\usage{
multitaper(
  ts,
  tapers,
  delta = 1,
  acf = FALSE,
  return_ff = TRUE,
  weights = NULL,
  ...
)
}
\arguments{
\item{ts}{A numeric vector representing the time series.}

\item{tapers}{A matrix of tapers (each column is a taper).}

\item{delta}{The sampling interval (default is 1).}

\item{acf}{Logical indicating whether to treat ts as the ACF. Setting to TRUE
returns the expected value of the multitaper estimate (default is FALSE).}

\item{return_ff}{Logical indicating whether to return the frequency vector
(default is TRUE).}

\item{weights}{weighted values of each multitaper. Defaults to equal
weighting.}

\item{...}{Frequency subsetting options (all default to TRUE).
\code{incl_boundaries} Logical indicating whether to include the boundary
frequencies at zero and Nyquist. \code{one_sided} Logical indicating whether
to generate a one-sided frequency vector. \code{positive_freqs} Logical
indicating whether to include results as positive frequencies or
zero-centred.}
}
\value{
A list with components \code{ff} (the frequency vector) and \code{mt}
(the multitaper spectral estimate).
}
\description{
This function computes the multitaper spectral estimate of a time series
\code{ts} using the given set of tapers.
}
