% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_locations}
\alias{subset_locations}
\title{Subset frequency locations based on specified criteria.}
\usage{
subset_locations(x, ...)
}
\arguments{
\item{x}{A numeric vector with default R frequency orderings.}

\item{...}{Frequency subsetting options (all default to TRUE).
\code{incl_boundaries} Logical indicating whether to include the boundary
frequencies at zero and Nyquist. \code{one_sided} Logical indicating whether
to generate a one-sided frequency vector. \code{positive_freqs} Logical
indicating whether to include results as positive frequencies or
zero-centred.}
}
\value{
A numeric vector of subsetted frequency orderings.
}
\description{
This function subsets frequency locations based on the specified criteria,
such as whether to include boundaries, whether to consider only positive
frequencies, and whether to return a one-sided frequency vector.
}
\examples{
x <- seq(-0.5, 0.5, by = 0.1)
subset_locations(x)

}
